[Fix] Merge models & types
