+++ packages/sdk/tsconfig.json
@@ -17,9 +17,9 @@
         // Ensure that .d.ts files are created by tsc, but not .js files
         "declaration": true,
         "sourceMap": true,
-        "paths": { "@connectorShared/*": ["../connectors/src/external/*"] }
+        "paths": { "@connectorShared/*": ["../connectors/src/*"] }
     },
-    "include": ["src/**/*", "../connectors/src/external/*"],
+    "include": ["src/**/*", "../connectors/src/*"],
     "exclude": ["node_modules", "build"], // *** The files to not type check ***
     "typedocOptions": {
         "entryPoints": [
